More widget deletion fixes
Make sure that the widgets that have custom cleanup procedure (TableRow, MenuBar, ...)
get clean up on Destroy(),  even if the destroy was not called on the widget itself,
but on its embedded widget.
Also protect Image, FileChooser and TextBuffer from accessing null ptr in C.
And protect FileChooser from trying to delete strings twice when Destroy() is called twice.